// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const RoleType = {
  ADMIN: 'ADMIN',
  USER: 'USER',
  SUPER_ADMIN: 'SUPER_ADMIN',
  VIEWER_ADMIN: 'VIEWER_ADMIN',
} as const;

export type RoleType = (typeof RoleType)[keyof typeof RoleType];

export const PaymentType = {
  CASH: 'CASH',
  CARD: 'CARD',
} as const;

export type PaymentType = (typeof PaymentType)[keyof typeof PaymentType];

export const MeasureType = {
  HOUR: 'HOUR',
  DAY: 'DAY',
} as const;

export type MeasureType = (typeof MeasureType)[keyof typeof MeasureType];

export const StatusType = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export type StatusType = (typeof StatusType)[keyof typeof StatusType];

export const LevelType = {
  BASIC: 'BASIC',
  INTERMEDIATE: 'INTERMEDIATE',
  ADVANCED: 'ADVANCED',
} as const;

export type LevelType = (typeof LevelType)[keyof typeof LevelType];